requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.4
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0

Werkzeug>=3.0
pytesseract>=0.3.10
pdfplumber>=0.11
pdf2image>=1.17
Pillow>=10.0
scikit-learn>=1.3
numpy>=1.24
pdfminer.six>=20221105